Add max_threads (when missing) to defined CPUs under /shared/hardware:
We introduced this because currently, using quattor's get_num_of_cores_max_threads would return wrong values if max_threads
is not defined.

This also assumes that hyperthreading is always enabled and available, if it is not available we are setting
the max threads count to cores count.

This change was applied in 2 steps:

1. Add the entry where it does not exist and set it to 2x cores count.

find /shared/hardware/cpu/ -type f -exec sh -c '
  if ! grep -q "\"max_threads\"\\s*=\\s*" "$1"; then
    awk '"'"'{ if ($0 ~ /\"cores\"\s*=\s*[0-9]+\s*;/) {
      cores_value = $3;
      gsub(/;$/, "", cores_value);
      print $0;
      print "\"max_threads\" = " cores_value * 2 ";"
    } else {
      print
    }}'"'"' "$1" > tmp && mv tmp "$1"
  fi
' sh {} \;

2. Change the entry to make max_threads equal to cores count when hyperthreading is not available

Which is the case for:

pentium_XXXX
core_duo_XXXX
woodcrest_XXXX
And all the AMD CPUs

find . -type f -exec sh -c '
>       if grep -qE "pentium|core_duo|woodcrest|AMD" "$1"; then
>         sed -i -e "/\"cores\"\\s*=\\s*\\([0-9]\\+\\)\\s*;/h" \
>                -e "/\"max_threads\"\\s*=\\s*[0-9]\\+\\s*;/!b" \
>                -e "g" \
>                -e "s/\"cores\"\\s*=\\s*\\([0-9]\\+\\)\\s*;/\"max_threads\" = \\1;/" "$1"
>       fi
\;>     ' sh {} \;